pyproject.toml
[project]
name = "webcam-iot-ingest"
version = "0.1.0"
description = "API to ingest webcam images to object storage & send notifications"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aioboto3>=15.5.0",
    "fastapi>=0.135.1",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "isodate>=0.7.2",
    "paho-mqtt>=2.1.0",
    "pillow>=12.1.1",
    "prometheus-client>=0.24.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
